move deployer test mock to point of use
